Use cgo for netlink socket creation
When creating a netlink socket in a namespace, the Go runtime
may decide to create a thread to schedule a goroutine - for
instance, when a syscall is done. The new thread may inherit
the namespace of the first thread which may not be the expected
behaviour. Using cgo for these syscalls do not trigger the Go
runtime and therefore avoid this kind of bugs.